refactor: findDefaultVersionForCRD
