+++ src/redux/thunks/cluster/index.ts
@@ -1,3 +1,4 @@
 export {loadClusterResources, reloadClusterResources} from './loadClusterResources';
 export {startClusterConnection} from './startClusterConnection';
 export {stopClusterConnection} from './stopClusterConnection';
+export {findDefaultVersionForCRD} from './findDefaultVersionForCRD';
