+++ src/redux/thunks/cluster/loadClusterResources.ts
@@ -25,6 +25,8 @@ import {RootState} from '@shared/models/rootState';
 import {createKubeClient} from '@shared/utils/kubeclient';
 import {trackEvent} from '@shared/utils/telemetry';
 
+import {findDefaultVersionForCRD} from './findDefaultVersionForCRD';
+
 const getNonCustomClusterObjects = async (kc: any, namespace?: string, allNamespaces?: boolean) => {
   return Promise.allSettled(
     getRegisteredKindHandlers()
@@ -204,59 +206,6 @@ export const reloadClusterResources = createAsyncThunk<
 >('main/reloadClusterResources', loadClusterResourcesHandler);
 
 /**
- * Find the default version in line with the algorithm described at
- * https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definition-versioning/#version-priority
- */
-
-const crdVersionRegex = /(v)(\d*)(alpha|beta)?(\d*)?/;
-
-export function findDefaultVersion(crd: any) {
-  if (!crd?.spec?.versions) {
-    return undefined;
-  }
-
-  const versionNames: string[] = crd.spec.versions.map((v: any) => v.name);
-
-  versionNames.sort((a, b) => {
-    const m1 = crdVersionRegex.exec(a);
-    const m2 = crdVersionRegex.exec(b);
-
-    // do both versions match the regex?
-    if (m1 && m2) {
-      // do both have initial version number?
-      if (m1[2] && m2[2]) {
-        // is the initial version the same?
-        if (m1[2] === m2[2]) {
-          // do both have an alpha or beta tag?
-          if (m1[3] && m2[3]) {
-            // is the tag the same?
-            if (m1[3] === m2[3]) {
-              // do both have an alpha or beta version?
-              if (m1[4] && m2[4]) {
-                return parseInt(m1[4], 10) - parseInt(m2[4], 10);
-              }
-              return m1[4] ? 1 : -1;
-            }
-            // compare tags (negate for beta > alpha)
-            return -m1[3].localeCompare(m2[3]);
-          }
-          return m1[3] ? 1 : -1;
-        }
-        // compare version numbers
-        return parseInt(m2[2], 10) - parseInt(m1[2], 10);
-      }
-      return m1[2] ? 1 : -1;
-    }
-    if (m1) return 1;
-    if (m2) return -1;
-
-    return a.localeCompare(b);
-  });
-
-  return versionNames.length > 0 ? versionNames[0] : undefined;
-}
-
-/**
  * Load custom resource objects for CRDs found in cluster
  */
 
@@ -288,7 +237,7 @@ async function loadCustomResourceObjects(
         }
 
         // retrieve objects using latest version name
-        const version = findDefaultVersion(crd.object) || 'v1';
+        const version = findDefaultVersionForCRD(crd.object) || 'v1';
         return namespace
           ? k8sApi
               .listNamespacedCustomObject(crd.object.spec.group, version, namespace, crd.object.spec.names.plural)
